# openapi.yaml
openapi: 3.0.0
info:
  title: "{{title}}"
  version: "1.0.0"


# Avoid having a definitive base path here. Set the path in the actual paths - facilitate versions
# Example v1.0.0/login and v1.0.2 can both be specified
servers:
  - url: http://127.0.0.1:5003/
    description: relative path example
paths:
...
  /users/v1/{user_id}:

    get:
      summary: Returns the current users's entity object
      tags:
        - Users
      description: >

        Returns a user entity object from a valid user id

        Required Headers:

            Authorization request header

            Bearer Valid Token

        Errors:

            'token-invalid', 401
            'authorisation-required', 401

            'not-found', 404
            'forbidden', 403

      operationId: users.v1.endpoints.get_user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Returns a data object containing a user's entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdate'

    delete:
      summary: Deletes a user's account
      tags:
        - Users
      description: >

        Required Headers:

          Authorization request header

            Bearer Valid Token

        Errors:

            'token-invalid', 401
            'authorisation-required', 401

            'not-found', 404
            'forbidden', 403

      operationId: users.v1.endpoints.delete_account
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Returns status ‘ok’
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityIdResponse'

    put:
      summary: Update a user's entity
      tags:
      - Users
      description: >


        Updates a user's entity


        Required Headers:


            Authorization request header

            Bearer Valid Token


        Body Parameters:

            At least one property from the Model - See Model

        Errors:

            'token-invalid', 401
            'authorisation-required', 401

            'user-not-found', 404
            'forbidden', 403
            'email-invalid', 400
            'signup-code-invalid', 400

      operationId: users.v1.endpoints.update
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        description: Update Data
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Returns a data object containing a user's entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUpdate'

#  /users/v1:
#    get:
#      summary: Attempt to retrieve a list of user entities - Requires Admin Permissions!
#      tags:
#      - Users
#      description: >
#
#
#        Returns a list of user entity objects
#
#
#        Required Headers:
#
#
#            Authorization request header
#
#            Bearer Valid Admin Access Token
#
#        Errors:
#
#            'token-invalid', 401
#            'authorisation-required', 401
#
#            'users-not-found', 404
#
#      operationId: users.v1.endpoints.list_users
#      responses:
#        '200':
#          description: Returns a data object containing a list of user entities and next_page object with the next_page to get or None
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/UserListResponse'
...
  /users/v1/email_verification:
    get:
      summary: Verifies user's email
      tags:
        - Users
      description: >

          Verifies user's email using an email-token generated when signing-up the user

          Errors:

              'token-invalid', 401
              'authorisation-required', 401

              'user-not-found', 404

      operationId: users.v1.endpoints.email_verification
      parameters:
        - name: token
          description: Email verification token generated at sign-up time
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns OK


#  /users/v1/password_forgotten:
#    get:
#      summary: Sends an email with new password token to user
#      tags:
#        - Users
#      description: >
#
#          Sends a reset password  email with new password token to user.
#
#          Errors:
#
#              'user-not-found', 404
#
#
#      operationId: users.v1.endpoints.forgotten_password
#      parameters:
#        - name: email
#          in: query
#          description: User's email
#          type: string
#          format: query
#          required: true
#      responses:
#        '200':
#          description: Returns OK
#          content:
#            application/json:
#              schema:
#                 $ref: '#/components/schemas/SuccessResponse'
#
#      "security": []

#  /users/v1/password_change:
#    put:
#      summary: Changes a users's password
#      tags:
#        - Users
#
#      description: >
#
#        Required Headers:
#
#            Authorization request header
#
#              Bearer Valid password token
#
#
#        Required Body Parameters:
#
#            password: New user password
#            email: User's email address
#
#
#        Errors:
#
#            'token-invalid', 401
#            'authorisation-required', 401
#
#            'password-invalid', 400
#            'user-not-found', 404
#
#      operationId: users.v1.endpoints.change_password
#      parameters:
#        - in: body
#          name: auth
#          description: New user's password
#          schema:
#            $ref: '#/components/schemas/Credentials'
#
#      responses:
#        '200':
#          description: Returns OK
#          content:
#            application/json:
#              schema:
#                 $ref: '#/components/schemas/SuccessResponse'

#  /users/v1/password_reset/{user_id}:
#    put:
#      summary: Reset a users's password
#      tags:
#        - Users
#      description: >
#
#        Required Headers:
#
#            Authorization request header
#
#              Bearer Valid Token
#
#        Required URL Parameters:
#
#            old_password: The user's old password
#            new_password: The user's new password
#
#        Errors:
#
#            'token-invalid', 401
#            'authorisation-required', 401
#
#            'password-invalid', 400
#            'user-not-found', 404
#
#
#      operationId: users.v1.endpoints.reset_password
#      parameters:
#        - in: path
#          name: user_id
#          required: true
#          type: integer
#        - in: body
#          name: resetpassword
#          description: User's old and new passwords
#          schema:
#            $ref: '#/components/schemas/ResetPassword'
#      responses:
#        '200':
#          description: Returns OK
#          content:
#            application/json:
#              schema:
#                 $ref: '#/components/schemas/SuccessResponse'

#  /users/v1/resend_email_verification/{user_id}:
#    put:
#      summary: Reset an email verification email to a user
#      tags:
#        - Users
#      description: >
#
#        Required Headers:
#
#            Authorization request header
#
#              Bearer Valid Token
#
#        Errors:
#
#            'token-invalid', 401
#            'authorisation-required', 401
#
#            'user-not-found', 404
#
#      operationId: users.v1.endpoints.resend_email_verification
#      parameters:
#      - in: path
#        name: user_id
#        required: true
#        schema:
#          type: integer
#      responses:
#        '200':
#          description: Returns OK
#          content:
#            application/json:
#              schema:
#                 $ref: '#/components/schemas/SuccessResponse'
  /users/v1/generate_tokens:
    get:
      summary: Generates new user access and refresh tokens
      tags:
        - Users
      description: >

        Required Headers:

          Authorization request header

            Bearer Valid Refresh Token

        Errors:

          'token-invalid', 401
          'authorisation-required', 401
          'user-not-found', 404

      operationId: users.v1.endpoints.generate_new_tokens

      responses:
        '200':
          description: Returns a new access token (token) and a new refresh token (refresh_token)
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/UserTokens'

      security:
        - jwt: []

# -----------------------------------------------
# COMPONENTS
# -----------------------------------------------
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: auth.endpoints.decode_token

  # Common Parameters
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/StatusOk'
    StatusOk:
      type: string
      description: Api call success
      default: ok
...
    EntityIdResponse:
      properties:
        id:
          type: integer
          description: Object's ID
    Email:
      description: Email Address
      type: string
      pattern: ([-!#-'*+/-9=?A-Z^-~]+(\.[-!#-'*+/-9=?A-Z^-~]+)*|"([]!#-[^-~ \t]|(\\[\t -~]))+")@[0-9A-Za-z]([0-9A-Za-z-]{0,61}[0-9A-Za-z])?(\.[0-9A-Za-z]([0-9A-Za-z-]{0,61}[0-9A-Za-z])?)+
    Password:
      description: Password
      type: string
      pattern: (?=\S{8,32})(?=\S*[A-Z])(?=\S*[a-z])(?=\S*[0-9])(?=\S*)(?<!\S)\S{8,32}(?=\s|\Z)


# -----------------------------------------------
#  CHARACTER SCHEMAS
# -----------------------------------------------

    # -----------------------------------------------
    #  Character DATA SCHEMAS
    # -----------------------------------------------
...
    BaseUser:
      type: object
      properties:
        email:
          $ref: '#/components/schemas/Email'

    # -----------------------------------------------
    #  AUTH TOKEN RESPONSE SCHEMAS
    # -----------------------------------------------
    UserTokens:
      type: object
      properties:
        token:
          type: string
          format: byte
          description: User's api calls token
        refresh_token:
          type: string
          format: byte
          description: User's refresh token

#    TokensResponse:
#      type: object
#      properties:
#        tokens:
#          $ref: '#/components/schemas/UserTokens'
#
#    EmailTokenResponse:
#      type: object
#      properties:
#        token:
#          type: string
#          description: An email token
#
#    PasswordTokenResponse:
#      type: object
#      properties:
#        token:
#          type: string
#          description: An email token

    # -----------------------------------------------
    #  USER AND USER AUTH SCHEMAS
    # -----------------------------------------------
...
    UserUpdate:
      type: object
      allOf:
        - $ref: '#/components/schemas/BaseUser'
        - $ref: '#/components/schemas/AccessRole'
...
    AccessRole:
      properties:
        access_role:
          description: Access role of user
          type: string
          enum:
            - admin
            - basic
          default: basic
    Credentials:
      type: object
      required:
        - password
        - email
      properties:
        password:
          $ref: '#/components/schemas/Password'
        email:
          $ref: '#/components/schemas/Email'

#    ResetPassword:
#      type: object
#      required:
#        - old_password
#        - new_password
#      properties:
#        old_password:
#          $ref: '#/components/schemas/Password'
#        new_password:
#          $ref: '#/components/schemas/Password'